+++ src/helper/types.ts
@@ -13,6 +13,7 @@ export type HpTrackerMetadata = {
     stats: {
         initiativeBonus: number;
         tempHp?: number;
+        initial?: boolean;
     };
     ruleset?: Ruleset;
     index?: number;
@@ -96,3 +97,11 @@ export type ACItemChanges = {
 };
 
 export type Ruleset = "e5" | "pf";
+
+export type InitialStatblockData = {
+    hp: number;
+    ac: number;
+    bonus: number;
+    slug: string;
+    ruleset: Ruleset;
+};